+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### Unreleased
+
+- Add RowsOfFieldsWithMetadata: allows commands to return an object with metadata that shows up in yaml/json (& etc.) formats, but is not shown in table/csv (& etc.).
+- Add NumericCellRenderer: allows commands to attach a renderer that will right-justify and add commas to numbers in a column.
+
 ### 3.1.13 - 29 November 2017
 
 - Allow XML output for RowsOfFields (#60).